moved currency logic to OrderService; removed obsolete AppService
