+++ CKODemoShop/ClientApp/src/app/app.component.ts
@@ -1,8 +1,8 @@
 import { Component, OnInit, OnDestroy } from '@angular/core';
 import { ICurrency } from './interfaces/currency.interface';
-import { AppService } from './services/app.service';
 import { FormBuilder, FormGroup, Validators } from '@angular/forms';
 import { Subscription } from 'rxjs';
+import { OrderService } from './services/order.service';
 
 @Component({
   selector: 'app-root',
@@ -12,11 +12,11 @@ import { Subscription } from 'rxjs';
 export class AppComponent implements OnInit, OnDestroy {
   subscriptions: Subscription[] = [];
   title: string = 'CKO Demo';
-  currencies: ICurrency[] = this._appService.currencies;
+  currencies: ICurrency[] = this._orderService.currencies;
   selectedCurrency: ICurrency;
   currencyForm: FormGroup;
 
-  constructor(private _appService: AppService, private _formBuilder: FormBuilder) { }
+  constructor(private _orderService: OrderService, private _formBuilder: FormBuilder) { }
 
   ngOnInit() {
     this.selectedCurrency = this.currencies[0];
@@ -24,8 +24,8 @@ export class AppComponent implements OnInit, OnDestroy {
       currency: [this.selectedCurrency, Validators.required]
     });
     this.subscriptions.push(
-      this._appService.currency$.subscribe(currency => this.selectedCurrency = currency),
-      this.currencyForm.get('currency').valueChanges.subscribe(currency => this._appService.setCurrency(currency)) // TODO: debug selection of currency
+      this._orderService.currency$.subscribe(currency => this.selectedCurrency = currency),
+      this.currencyForm.get('currency').valueChanges.subscribe(currency => this._orderService.setCurrency(currency)) // TODO: debug selection of currency
     );
   }
